+++ SqExpress/SqQueryBuilder.cs
@@ -27,8 +27,6 @@ namespace SqExpress
 
         public static ExprUnsafeValue UnsafeValue(string unsafeValueExpr) => new ExprUnsafeValue(unsafeValueExpr);
 
-        public static ExprUnsafeQuery UnsafeQuery(string unsafeQueryExpr) => new ExprUnsafeQuery(unsafeQueryExpr);
-
         public static ExprIsNull IsNull(ExprValue value) => new ExprIsNull(value, not: false);
 
         public static ExprIsNull IsNotNull(ExprValue value) => new ExprIsNull(value, not: true);